+++ addons/ueipab_attendance_report/models/__init__.py
@@ -3,3 +3,4 @@ from . import hr_attendance_correction
 from . import hr_notice_acknowledgment
 from . import partner_communication_ack
 from . import hr_leave
+from . import hr_employee_kiosk_guard